// client/src/utils/type.ts
export interface User {
  id: string;
  email: string;
  name: string;
  imageUrl: string | null;
}
// ...
export interface Message {
  id: string;
  body: string;
  conversationId: string;
  senderId: string;
  createdAt: string;
  updatedAt: string;
  sender: Sender;
  fileId?: string | null;
  fileUrl?: string | null;
}

export interface Sender {
  id: string;
  joinedAt: string;
  userId: string;
  conversationId: string;
  role: string;
  user: User;
}
// Generated by https://quicktype.io

export interface Conversation {
  id: string;
  isGroup: boolean;
  groupTitle: null;
  type: string;
  createdAt: string;
  updatedAt: string;
  members: Member[];
}

export interface Member {
  id: string;
  joinedAt: string;
  userId: string;
  conversationId: string;
  role: string;
  user: User;
}
// ...
export interface ChatListItemProps {
  conversation: Conversation;
  currentConversation: Conversation | null;
  newMessagesInConversations: Message[];
  handleResetNewMessagesInConversation: (cId: string) => void;
}